django_dashboard/services/analytics_service.py
"""
Analytics Service for MathSprout
==============================

This module provides analytics functionality for the MathSprout application,
including data aggregation, analysis, and visualization capabilities.
"""

import os
import pandas as pd
import numpy as np
from datetime import datetime, timedelta
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from dotenv import load_dotenv
from .supabase_service import SupabaseService


class AnalyticsService:
    def __init__(self):
        """Initialize the analytics service with optional Supabase connection."""
        # Load environment variables
        load_dotenv()

        # Debug: Print current working directory and environment variables
        print(f"Current working directory: {os.getcwd()}")
        print(f"Environment variables loaded: {os.getenv('SUPABASE_URL') is not None}")

        # Try to initialize Supabase connection
        supabase_url = os.getenv("SUPABASE_URL")
        supabase_key = os.getenv("SUPABASE_KEY")

        print(f"Supabase URL found: {supabase_url is not None}")
        print(f"Supabase Key found: {supabase_key is not None}")

        self.use_mock_data = not (supabase_url and supabase_key)
        if not self.use_mock_data:
            print("Initializing Supabase connection...")
            self.supabase = SupabaseService()
            self.supabase.initialize_connection(supabase_url, supabase_key)
        else:
            print(
                "Warning: No Supabase credentials found. Using mock data for development."
            )

    def get_classroom_analytics(self, classroom_id: str) -> dict:
        """Get analytics for a specific classroom."""
        if self.use_mock_data:
            return self._generate_mock_classroom_data()

        # Real implementation would go here
        pass

    def get_student_analytics(self, student_id: str) -> dict:
        """Get analytics for a specific student."""
        if self.use_mock_data:
            return self._generate_mock_student_data()

        # Real implementation would go here
        pass

    def get_performance_trends(self, time_period: str) -> dict:
        """Get performance trends for the specified time period."""
        if self.use_mock_data:
            return self._generate_mock_trend_data(time_period)

        # Real implementation would go here
        pass

    def _generate_mock_classroom_data(self) -> dict:
        """Generate mock classroom data for development."""
        np.random.seed(42)  # For reproducible mock data

        # Generate mock score distribution
        scores = np.random.normal(75, 15, 100)
        hist, edges = np.histogram(scores, bins=10)

        # Generate mock skill data
        skills = ["Addition", "Subtraction", "Number Recognition", "Counting"]
        skill_means = {skill: np.random.uniform(60, 90) for skill in skills}
        skill_std = {skill: np.random.uniform(5, 15) for skill in skills}

        # Generate mock cluster data
        n_students = 50
        pca_components = [
            np.random.normal(0, 1, n_students),
            np.random.normal(0, 1, n_students),
        ]
        cluster_assignments = KMeans(n_clusters=3).fit_predict(
            np.array(pca_components).T
        )

        return {
            "mean_performance": {
                "mean_score": float(np.mean(scores)),
                "median_score": float(np.median(scores)),
                "std_dev": float(np.std(scores)),
            },
            "overall_performance": {
                "score_distribution": {"bins": hist.tolist(), "edges": edges.tolist()}
            },
            "skill_distribution": {"skill_means": skill_means, "skill_std": skill_std},
            "student_clusters": {
                "pca_components": pca_components,
                "cluster_assignments": cluster_assignments.tolist(),
            },
            "performance_trends": {
                "trend_direction": (
                    "increasing" if np.random.random() > 0.5 else "decreasing"
                )
            },
        }

    def _generate_mock_student_data(self) -> dict:
        """Generate mock student data for development."""
        np.random.seed(42)  # For reproducible mock data

        # Generate mock timeline data
        dates = [datetime.now() - timedelta(days=x) for x in range(30)]
        base_score = 70
        noise = np.random.normal(0, 5, len(dates))
        trend = np.linspace(0, 15, len(dates))  # Upward trend
        scores = base_score + noise + trend

        # Generate mock skill data
        skills = ["Addition", "Subtraction", "Number Recognition", "Counting"]
        skill_scores = np.random.uniform(60, 90, len(skills))

        # Generate mock game performance data
        game_types = ["Speed Math", "Memory Match", "Number Line", "Count & Sort"]
        game_scores = {game: np.random.uniform(70, 90) for game in game_types}

        return {
            "performance_summary": {"current_score": float(scores[-1])},
            "learning_patterns": {
                "learning_rate": float(np.random.uniform(0.5, 2.0)),
                "consistency_score": float(np.random.uniform(0.6, 0.9)),
            },
            "progress_timeline": {
                "dates": [d.strftime("%Y-%m-%d") for d in dates],
                "scores": scores.tolist(),
            },
            "skill_breakdown": {"skills": skills, "scores": skill_scores.tolist()},
            "game_type_performance": game_scores,
        }

    def _generate_mock_trend_data(self, time_period: str) -> dict:
        """Generate mock trend data for development."""
        np.random.seed(42)  # For reproducible mock data

        # Generate dates based on time period
        if time_period == "Last Week":
            days = 7
        elif time_period == "Last Month":
            days = 30
        else:  # Last Quarter
            days = 90

        dates = [datetime.now() - timedelta(days=x) for x in range(days)]
        dates.reverse()

        # Generate mock rolling average and volatility
        base_trend = np.linspace(70, 85, len(dates))
        noise = np.random.normal(0, 3, len(dates))
        rolling_avg = base_trend + noise
        volatility = np.abs(np.random.normal(5, 2, len(dates)))

        # Generate mock heatmap data
        heatmap = np.random.uniform(60, 90, size=(5, 3))  # 5 days x 3 time periods

        # Generate mock improvement areas
        improvement_areas = [
            {"name": "Problem Solving Speed", "progress": 0.75},
            {"name": "Accuracy", "progress": 0.85},
            {"name": "Concept Understanding", "progress": 0.60},
        ]

        return {
            "dates": [d.strftime("%Y-%m-%d") for d in dates],
            "rolling_average": rolling_avg.tolist(),
            "volatility": volatility.tolist(),
            "heatmap": heatmap.tolist(),
            "improvement_areas": improvement_areas,
        }
