requirements.txt
Django==4.2.20
python-dotenv==1.0.1
supabase==2.3.4
httpx>=0.24,<0.26
gotrue==2.5.5
pandas
numpy
scikit-learn
plotly
absl-py==2.2.2
altair==4.1.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bokeh==3.4.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cloudscraper==1.2.71
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
gymnasium==1.1.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
# -e git+https://github.com/lmorgenthaler/MathSprout.git@eedf5f78b55e1ef434ad269610f00a2fca4efa59#egg=mathsprout
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mizani==0.11.4
mpmath==1.3.0
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.2.1
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
patsy==1.0.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
plotnine==0.13.6
postgrest==0.15.1
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
realtime==1.0.6
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
SciencePlots==2.1.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.31.0
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
storage3==0.7.7
StrEnum==0.4.15
supafunc==0.3.3
swig==4.3.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0
xyzservices==2025.4.0
zipp==3.21.0

